User Sign Up
Did the front end of the user sign up as a modal, because we don’t
really need a whole page for the signup part. Also converted the style
sheet to bootstrap settings. We can put all the style information from
csscreate file into the bootstrap css file later on, so that the
welcome.php is less cluttered.